Fix validation of code-owner.config files
The validation may be invoked for a commit which is not present in the
repository yet and trying to load it via a new RevWalk fails with
MissingObjectException. Such commits are only visible through the
RevWalk that is creating the commit. This RevWalk is passed in together
with the CommitReceivedEvent. We must use this RevWalk instance to load
the commit from the event.

For example the validation of code-owner.config files failed with
MissingObjectException when a change for refs/meta/config was rebased.
It failed regardless of whether the change actually touched a
code-owners.config file (because we need to load the commit to know if
the code-owners.config file was touched).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I5fc34f0302fbf77908fba56da5d94924a0d05bd8