Test NotImplemented Java API implementations
The test makes sure that all methods throw NotImplementedException.
Admittedly this test adds only very little value, but having it makes
the test coverage for the api package green (which is the only package
that isn't green yet) and hence safes me from checking what's wrong with
this package each time I'm looking at the test coverage. Since only
little test code is needed to cover the NotImplemented classes, I think
it's fine to have it. With this test the test coverage for the api
package increases from 81.3% to 100%.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I4007096c4bf2cb0dfb61e909cf72b17bb7945991