Consider only emails that belong to multiple active accounts as ambiguous
Code owner emails that belong to multiple accounts are detected as
ambiguous and their code ownership is ignored. This is intended however
if an email belongs to multiple accounts and only of them is active,
that's fine as there is exactly one active account to which the code
ownership is assigned.

For CodeOwnerResolver this means that it has to load the accounts, and
check if they are active, before deciding if an email is ambiguous.
Doing this changes, actually improves, the debug logs in case there is
an orphaned email.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I50f06e4e055ee04ac17c8873f0e08c0d6aca4b5e