Validation: Log exceptions that happen in dry run mode of as a warning
The dry run mode allows to test the validation and it should be possible
to check that it's working before turning it on for real. The idea is to
know whether it's working by checkng the server log for exceptions that
happen during the validation. Unfortunately these exceptions were only
logged on fine level, so that they do not show up in server logs by
default. Raise the log level to warning so that the exceptions are
always logged and the dry run mode can be used as intended.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I9ba4460995c7447b4c59a5a15774a5d3ff6cf67f