Do not run the code owners submit rule for closed changes
Running the code owners submit rule is rather expensive, since we need
to compute the code owner statuses for all files which have been touched
in the change, which requires loading the relevant OWNERS files for each
of the files. Submit rules are also invoked for closed changes (e.g.
each time a change is reindexed), but the code owner status on closed
changes is not relevant, hence we should not compute it.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I6060d8788cb7ba9136ad395ac8bffe402d8d5b5c