Allow to set a max size for the in-memory code owner config cache
If this cache is unlimited there is a danger of caching too many code
owner config files and causing out-of-memory errors.

For now we do not set any limit per default as we don't know yet what
would be a good default value (but change I40f1eaae0 adds a new metric
that should give us an idea about this).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I1837fc1addf522ce1c044c5c82ea77d843a034c4