Allow to exempt users from requiring code owner approvals
If a user is exempted from requiring code owner approvals changes that
are uploaded by this user are automatically code-owner approved.

This is useful to exempt bots that do automatic integrations.

Exempted users are specified by email in the code-owners.config or
gerrit.config. This is consistent with how global code owners are
specified. Exempted users that are specified by other account
identifiers (e.g. full name, account ID) are ignored.

Change-Id: I5d50ebddcdbf65ee31a0fd35ab89dc19c1f594e6
Signed-off-by: Edwin Kempin <ekempin@google.com>