Reduce default value for maxPathsInChangeMessages config parameter
The maxPathsInChangeMessages controls how many owned paths should be
included into changes messages at most. Computing owned path is rather
expensive and contributes to the latency of post review. By reducing the
number of owned paths that are included by default we lower the impact
on the latency. The paths in the change messages are used for having a
more transparent UX, but also for code owner metrics. Reducing the
number of the included paths makes the metrics less precise, but since
most changes touch less than 50 files the new default value should be
still good enough.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I3b520220c76e1f64ec7814f40f0e82d3c48ff0d4