Allow validating code owner configs even if code owners is switched off
When rolling out code-owners, it can make sense to enable the validation
of code owner config files before enabling the code owners
functionality. To support this we add 'FORCED' and 'FORCED_DRY_RUN' as
new options for the enableValidationOnCommitReceived and
enableValidationOnSubmit configuration parameters, since 'TRUE' and
'DRY_RUN' only trigger the validation if the code owners functionality
is enabled.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I29fdbd165e067bdab442f1d1d5e54dc537c67369