Disable saving auto-merge commits
If an auto-merge is not available yet, it needs to be created. Saving it
means doing a write in Git. The changed files computation is often done
in read requests, and doing a Git write in a read request is bad. Hence
it's better to disable saving auto-merges, even if it means that it
needs to be recomputed later again.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I2ecce2a86b8fd403e3c1dec8fab7dd301f21cb74