Test code-owners submit rule if auto-merge is missing
If the auto-merge is missing in Git it needs to be computed in memory
when the code-owners submit rule is executed for a merge commit and
mergeCommitStrategy is configured to FILES_WITH_CONFLICT_RESOLUTION.
This was not yet covered by a test and it was failing after change
I2ecce2a86 got submitted (because AutoMerger failed if InMemoryInserter
was passed in but change.cacheAutomerge was true and it attempted to
save the auto-merge). Cover this by a test now. In meantime this has
been fixed by change Iebb1a5a8d so that the test is passing now.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I21abfb90dbcf694607e1b766a92fd5dcb432eb11