Disable validation of OWNERS files on submit by default
Changes to OWNERS files are already validated on upload / commit
creation, hence there is no urgent need to validate them once again on
submit. This means we can skip the validation on submit by default.

There are reasons for validating OWNERS files on submit:

1. if the validation of OWNERS files on upload / commit creation is
   switched off (e.g. because it is preferred to report issues with
   OWNERS files as findings rather than rejecting the upload / commit
   creation)
2. if changes stay open for a long time it can happen that changes to
   OWNERS files that were valid on upload are invalid on submit (e.g. a
   newly added import was resolvable on upload, but then the imported
   OWNERS file got deleted and it's no longer available when the submit
   is done)

1. is not the case by default and 2. happens rarely enough so that by
default we do not need to do the validation on submit.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Iba334f1a92b7f0e556855982198de9a52eefada0