Show different error message if server config is invalid
If a code-owners operation cannot be performed because the project contains an invalid OWNERS file the backend returns a '409 Conflict' response informing about the invalid OWNERS file. An invalid OWNERS file is a configuration issue that needs to be fixed by the project owner or host admin, it's not an error in the code-owners plugin (an error would be signaled as 500 response). At the moment the frontend reports invalid OWNERS file as a red error banner (see screenshot in b/183641693#comment1). This gives the wrong impression that there was an error in the code-owners plugin. The way how invalid OWNERS files are reported/shown to the user is changed so that it's more clear that this is a configuration issue and not an error in the plugin.

Change-Id: I0e2524cbd1c76f954b77aff19a0da06b78574aec