GeneralConfigTest: Fix wrong config parameter name
The inheritedExemptedUsersCanBeRemovedOnProjectLevel test is supposed to
verify that exempted users that are configured in gerrit.config can be
removed. However since the config paramater name was wrong there were no
exempted users configured in gerrit.config and hence this test failed to
do the verification it was supposed to do.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Iedc46ae82e4a369f5582504c6cc00f1ab0e6bae2