CodeOwnersPluginConfigSnapshotTest: Add tests for global code owners
The tests are similar to the ones for exempted users.

We intent to change the way how inheritance works for list config
parameters, such as global code owners. With the new tests the changed
behavior will be covered and it will also make it easier for reviewers
to see what is changing.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I337f70d3b3cda2a82133e6e77aa4f71c9dcbb038