CodeOwnersPluginConfigSnapshotTest: Add more override approval tests
The tests are similar to the ones for required approvals.

We intent to change the way how inheritance works for list config
parameters, such as override approvals. With the new tests the changed
behavior will be covered and it will also make it easier for reviewers
to see what is changing.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I305e3b24a5cb2319a0715a4cf4ab7d71b322116f